Language.C.Simple.CValue

Documentation

data CValue

A generic C value

Constructors

VStruct [CValue]

VUnion UnionPath CValue

VPrimitive PrimitiveValue

VArray [CValue]

VMember CValue

Void

class ToCValue a where

ToCValue Class

Methods

toCValue :: a -> CValue

class FromCValue a where

Derive from this to convert from a CValue to Haskell type

Methods

fromCValue :: CValue -> Either String a

data Side

A step in union path

Constructors

Lft

Rght

type UnionPath = [Side]

This is used for the conversion from a CValue back to Haskell type. Ideally it should be and index, but unforunately this does not work with the way Generics creates its :+: binary tree. I'm leaving it here for now, but I might find a more elegant way to handle this.

data PrimitiveValue

Primitive C values

Constructors

PChar CChar

PSChar CSChar

PUChar CUChar

PShort CShort

PUShort CUShort

PInt CInt

PUInt CUInt

PLong CLong

PULong CULong

PPtrdiff CPtrdiff

PSize CSize

PWchar CWchar

PSigAtomic CSigAtomic

PLLong CLLong

PULLong CULLong

PIntPtr CIntPtr

PUIntPtr CUIntPtr

PIntMax CIntMax

PUIntMax CUIntMax

PClock CClock

PTime CTime

PUSeconds CUSeconds

PSUSeconds CSUSeconds

PFloat CFloat

PDouble CDouble